align node folder to cumulus parachain-template
