+++ node/src/main.rs
@@ -16,12 +16,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#![warn(missing_docs)]
+
 mod chain_spec;
-mod cli;
 #[macro_use]
 mod service;
+mod cli;
 mod command;
-mod executor;
 mod rpc;
 
 fn main() -> sc_cli::Result<()> {
